Check for valid password before actually executing archive deletion.
This commit also adds a new library file, `lib/FetLife.php`, which is a
simplified version of the FetLife Bridge plugin for StatusNet. For more
information about its source, see:

    https://github.com/meitar/fetlife-bridge/

Future functionality for interfacing with FetLife.com should be placed
in this library file, in which case it may also become a git submodule.